re-organize directories

+++ src/others/llvm-link/adder.h
@@ -1,1 +1,1 @@
-long int adder(int a, int b);
+long long unsigned int adder(int a, int b);
